pyproject.toml
[tool.poetry]
name = "fast_python-api"
version = "0.1.0"
description = "test_task"
authors = ["Asgef <asgefes1@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.1"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
aiohttp = "^3.11.11"
pydantic-settings = "^2.7.1"
sqlalchemy = "^2.0.37"
asyncpg = "^0.30.0"
alembic = "^1.14.0"
pyjwt = "^2.10.1"
sqids = "^0.5.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^2.10.6"}
python-multipart = "^0.0.20"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
aioresponses = "^0.7.7"
httpx = "^0.28.1"
pytest-asyncio = "^0.25.2"
pytest = "^8.3.4"
respx = "^0.22.0"
pytest-cov = "^6.0.0"
ipython = "^8.32.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
